test(usability): fix float parsing hallucination in stats tools limits
History-Category: Changed
History-Impact: 0.1
History-Confidence: 1
History-Validation: passed
History-Journal: true

+++ src/adapters/mysql/tools/stats/advanced.ts
@@ -103,7 +103,7 @@ export const StatsTopNSchema = z.preprocess(
     database: z.string().optional(),
     table: z.string().min(1, "table is required"),
     column: z.string().min(1, "column is required"),
-    n: z.number().min(1).max(100).default(10),
+    n: z.number().int().min(1).max(100).default(10),
     direction: z.enum(["asc", "desc"]).default("desc"),
     selectColumns: z.array(z.string()).optional(),
     where: z.string().optional(),
@@ -149,7 +149,7 @@ export const StatsDistinctSchema = z.preprocess(
     database: z.string().optional(),
     table: z.string().min(1, "table is required"),
     column: z.string().min(1, "column is required"),
-    limit: z.coerce.number().min(1).max(1000).default(100),
+    limit: z.coerce.number().int().min(1).max(1000).default(100),
     where: z.string().optional(),
   })
 );
@@ -196,7 +196,7 @@ export const StatsFrequencySchema = z.preprocess(
     database: z.string().optional(),
     table: z.string().min(1, "table is required"),
     column: z.string().min(1, "column is required"),
-    limit: z.coerce.number().min(1).max(1000).default(20),
+    limit: z.coerce.number().int().min(1).max(1000).default(20),
     where: z.string().optional(),
   })
 );
